model/requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
Pillow==9.5.0